Extract views into its own files
